+++ typescript.js
@@ -1,6 +1,19 @@
 module.exports = {
+    parser: '@typescript-eslint/parser',
+    parserOptions: {
+        project: './tsconfig.json',
+        ecmaVersion: 2020,
+        sourceType: 'module',
+    },
+    plugins: ['@typescript-eslint'],
     extends: [
-        'plugin:@typescript-eslint/recommended',
+        'plugin:import/typescript',
     ],
-    rules: {},
+    overrides: [{
+        files: ['*.ts', '*.tsx'],
+        extends: [
+            'plugin:@typescript-eslint/recommended',
+        ],
+        rules: {},
+    }],
 }