+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v0.3.0 / 2021-10-05
+===================
+- Updated few style rules
+- Updated `pronos/typescript` config
+
+v0.2.0 / -
+==========
+- Missing changelog
+
 0.1.0 / 2021-09-13
 ==================
 - Initial version of the config